feat(container metrics): change timeRange parameter type from Int to Long for better precision in time handling refactor(container metrics): rename ContainerMetricsPort to CpuMemoryMetricsPort for clarity and update related references feat(trace analysis): implement TraceAnalysisService to analyze HTTP requests, status codes, and latency from traces chore(trace repository): add TraceCountPerMinute data class to represent trace counts per minute refactor(span model): add methods to check if a span is an HTTP request and to retrieve its status code for better trace analysis

+++ src/main/kotlin/xquare/app/xquareinfra/adapter/in/container/V2ContainerWebAdapter.kt
@@ -94,7 +94,7 @@ class V2ContainerWebAdapter(
         @RequestParam("environment", required = true)
         environment: ContainerEnvironment,
         @RequestParam("timeRange", required = true)
-        timeRange: Int
+        timeRange: Long
     ): Map<String, Map<String, String>> =
         containerMetricUseCase.getContainerHttpRequestPerMinute(
             deployId,
@@ -110,7 +110,7 @@ class V2ContainerWebAdapter(
         @RequestParam("environment", required = true)
         environment: ContainerEnvironment,
         @RequestParam("timeRange", required = true)
-        timeRange: Int,
+        timeRange: Long,
         @PathVariable("statusCode", required = true) statusCode: Int
     ): Map<String, Map<String, String>> =
         containerMetricUseCase.getContainerHttpStatusRequestPerMinute(
@@ -129,7 +129,7 @@ class V2ContainerWebAdapter(
         environment: ContainerEnvironment,
         @PathVariable("percent", required = true) percent: Int,
         @RequestParam("timeRange", required = true)
-        timeRange: Int
+        timeRange: Long
     ): Map<String, Map<String, String>> =
         containerMetricUseCase.getContainerLatency(
             deployId,
